[iris] Merge Autoscaler + Scheduler dashboard tabs into one Capacity & Scheduling view
The two tabs told contradictory stories with no way to correlate them — the
Autoscaler tab could show a slice "idle — eligible for scale-down" while the
Scheduler tab said "waiting for v5ps". Replace both with a single pool-keyed
view that answers, in one place: what is scheduling / not scheduling, where &
why, and what is available.

CapacityTab.vue joins per-pool demand + availability + slice health + pending
jobs + user quotas. The headline element reads the new server-side usability
rollups: per-slice schedulable/degraded badges, and — crucially — a slice that
is idle but degraded renders "idle · unschedulable (degraded)" distinctly from a
plain reclaimable "idle", computed per-slice from SliceInfo.degradedSlotCount.
A per-pool note distinguishes "placeable capacity with unmet demand" (a real
bug) from "demand blocked on degraded workers" (recovery in progress).

Deletes AutoscalerTab.vue (1164) + SchedulerTab.vue (451), adds CapacityTab.vue
(~1176) — net frontend reduction. Drops the Fleet Overview region aggregation
and the dead per-job demand stub; the server-side rollups replace the old
client-side in-use joins. Routes /autoscaler + /scheduler collapse to /capacity;
the e2e smoke test follows.
